+++ TeachingRecordSystem/src/TeachingRecordSystem.Core/ApiSchema/V3/V20250804/WebhookData/AlertCreatedNotification.cs
@@ -1,19 +1,19 @@
 using TeachingRecordSystem.Core.ApiSchema.V3.V20240920.Dtos;
 
-namespace TeachingRecordSystem.Core.ApiSchema.V3.VNext.WebhookData;
+namespace TeachingRecordSystem.Core.ApiSchema.V3.V20250804.WebhookData;
 
-public record AlertDeletedNotification : IWebhookMessageData
+public record AlertCreatedNotification : IWebhookMessageData
 {
-    public static string CloudEventType { get; } = "alert.deleted";
+    public static string CloudEventType { get; } = "alert.created";
 
     public required string Trn { get; init; }
     public required Alert Alert { get; init; }
 }
 
-public class AlertDeletedNotificationMapper(PersonInfoCache personInfoCache, ReferenceDataCache referenceDataCache) :
-    IEventMapper<AlertDeletedEvent, AlertDeletedNotification>
+public class AlertCreatedNotificationMapper(PersonInfoCache personInfoCache, ReferenceDataCache referenceDataCache) :
+    IEventMapper<AlertCreatedEvent, AlertCreatedNotification>
 {
-    public async Task<AlertDeletedNotification?> MapEventAsync(AlertDeletedEvent @event)
+    public async Task<AlertCreatedNotification?> MapEventAsync(AlertCreatedEvent @event)
     {
         if (@event.Alert.AlertTypeId is not Guid alertTypeId)
         {